Implemented tracking SCREEN_VIEW event and refactors
