+++ app/src/main/java/com/example/newsapp/MainActivity.kt
@@ -2,6 +2,7 @@ package com.example.newsapp
 
 import android.content.res.Configuration
 import android.os.Bundle
+import android.util.Log
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.fillMaxSize
@@ -11,67 +12,50 @@ import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.navigation.NavType
-import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
+import androidx.navigation.NavHostController
 import androidx.navigation.compose.rememberNavController
-import androidx.navigation.navArgument
 import com.example.newsapp.ui.theme.NewsAppTheme
-import com.example.newsapp.view.InfoScreen
-import com.example.newsapp.view.NewsDetailsScreen
-import com.example.newsapp.view.NewsListScreen
+import com.example.newsapp.view.Navigation
+import com.example.newsapp.view.screens.NewsListScreen
 import com.example.newsapp.view.ScreenScaffoldWithMenu
-import com.example.newsapp.view.WebViewScreen
+import com.google.firebase.Firebase
+import com.google.firebase.analytics.FirebaseAnalytics
+import com.google.firebase.analytics.analytics
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
 class MainActivity : ComponentActivity() {
+    private lateinit var firebaseAnalytics: FirebaseAnalytics
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+
+        firebaseAnalytics = Firebase.analytics
+
         setContent {
             NewsAppTheme {
                 Surface(modifier = Modifier.fillMaxSize(), color = MaterialTheme.colorScheme.background) {
                     val navController = rememberNavController()
+                    setupScreenTracking(navController)
                     val scaffoldState = rememberScaffoldState()
                     ScreenScaffoldWithMenu(
                         navController = navController,
                         scaffoldState = scaffoldState,
                         scaffoldContent = {
-                            NavHost(navController = navController, startDestination = Routes.NEWS_LIST_SCREEN.id) {
-                                composable(Routes.NEWS_LIST_SCREEN.id) { NewsListScreen(navController) }
-                                composable(Routes.INFO_SCREEN.id) { InfoScreen(navController) }
-                                composable(
-                                    route = Routes.NEWS_DETAILS_SCREEN.id.plus("/{id}"),
-                                    arguments = listOf(
-                                        navArgument("id") {
-                                            type = NavType.IntType
-                                        }
-                                    )
-                                ) { navBackStackEntry ->
-                                    navBackStackEntry.arguments?.getInt("id")?.let { newsId ->
-                                        NewsDetailsScreen(
-                                            navController = navController,
-                                            newsId = newsId
-                                        )
-                                    }
-                                }
-                                composable(
-                                    route = Routes.WEB_VIEW_SCREEN.id.plus("/{url}"),
-                                    arguments = listOf(
-                                        navArgument("url") {
-                                            type = NavType.StringType
-                                        }
-                                    )
-                                ) {navBackStackEntry ->
-                                    navBackStackEntry.arguments?.getString("url")?.let { url ->
-                                        WebViewScreen(urlToArticle = url)
-                                    }
-                                }
-                            }
+                            Navigation(navController = navController)
                         }
                     )
                 }
             }
+        }
+    }
+
+    private fun setupScreenTracking(navController: NavHostController) {
+        navController.addOnDestinationChangedListener { _, destination, _ ->
+            val params = Bundle()
+            params.putString(FirebaseAnalytics.Param.SCREEN_NAME, destination.route)
+            params.putString(FirebaseAnalytics.Param.SCREEN_CLASS, destination.route)
+            firebaseAnalytics.logEvent(FirebaseAnalytics.Event.SCREEN_VIEW, params)
+            Log.d("ScreenEvent", params.toString())
         }
     }
 }
